requirements.txt
python-telegram-bot==21.3
apscheduler==3.10.4
google-generativeai==0.7.2
python-dotenv==1.0.1
